Imp: Run execute callback in thread
We have seen some cases where the
execute_callback has an uncatchable failure
which makes the executeLoop and thereby
self.execute_thread crash. This goes unrecognized
by the client as the server keeps publising the
latest status to the status topic. With this change
we create another thread to isolate the execution
of the execute_callback, so that we can still abort
the goal if even if this new thread crashes.