Imp: Catch BaseException in SimpleActionServer
We have seen some cases where the
execute_callback has an uncaught failure
which makes the executeLoop and thereby
self.execute_thread terminate. This goes unrecognized
by the client as the server keeps publising the
latest status to the status topic. With this change
we extend the handling of Exceptions to all
BaseExceptions that are not inherited from the
Exception class.